Etapa 3: Resolví los problemas de falta de abstracción y reutilización mediante la aplicación de los principios 4, 5 y 6.

+++ practica-principios-diseno/clinicasegura/aplicacion/configuracion.py
@@ -1,7 +1,8 @@
 from dataclasses import dataclass
+from decimal import Decimal
 
 @dataclass(frozen=True)
 class Configuracion:
     vigencia_dias: int
-    tarifa_diaria: float
+    tarifa_diaria: Decimal
     timeout: float
